requirements.txt
anthropic
python-dotenv
playwright
